# NO-TICKET: Optimize FromBytes for Vec<u8> case.