@henrytill 2020-01-03
My numbers against current dev:
```
Benchmarking tps/transfer_to_existing_account_multiple_execs/3: Warming up for 3.0000 s
Warning: Unable to complete 10 samples in 5.0s. You may wish to increase target time to 6.0s or reduce sample count to 10
tps/transfer_to_existing_account_multiple_execs/3                                                                          
                        time:   [109.01 ms 111.53 ms 113.32 ms]
                        thrpt:  [26.473  elem/s 26.899  elem/s 27.521  elem/s]
                 change:
                        time:   [-7.5946% -5.4948% -3.2697%] (p = 0.00 < 0.05)
                        thrpt:  [+3.3803% +5.8142% +8.2187%]
                        Performance has improved.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high mild
Benchmarking tps/transfer_to_existing_account_multiple_deploys_per_exec/3: Warming up for 3.0000 s
Warning: Unable to complete 10 samples in 5.0s. You may wish to increase target time to 5.7s or reduce sample count to 10
Benchmarking tps/transfer_to_existing_account_multiple_deploys_per_exec/3: Collecting 10 samples in estimated 5.6821 s (55 iteratio                                                                                                                                   tps/transfer_to_existing_account_multiple_deploys_per_exec/3                        
                        time:   [103.17 ms 103.92 ms 104.56 ms]
                        thrpt:  [28.691  elem/s 28.869  elem/s 29.078  elem/s]
                 change:
                        time:   [-9.3038% -8.4940% -7.7166%] (p = 0.00 < 0.05)
                        thrpt:  [+8.3618% +9.2825% +10.258%]
                        Performance has improved.
Benchmarking tps/transfer_to_purse_multiple_execs/3: Warming up for 3.0000 s
Warning: Unable to complete 10 samples in 5.0s. You may wish to increase target time to 6.2s or reduce sample count to 10
tps/transfer_to_purse_multiple_execs/3                                                                          
                        time:   [113.88 ms 115.13 ms 117.07 ms]
                        thrpt:  [25.625  elem/s 26.058  elem/s 26.344  elem/s]
                 change:
                        time:   [-7.3208% -4.9022% -2.4357%] (p = 0.00 < 0.05)
                        thrpt:  [+2.4965% +5.1549% +7.8991%]
                        Performance has improved.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high mild
Benchmarking tps/transfer_to_purse_multiple_deploys_per_exec/3: Warming up for 3.0000 s
Warning: Unable to complete 10 samples in 5.0s. You may wish to increase target time to 5.7s or reduce sample count to 10
tps/transfer_to_purse_multiple_deploys_per_exec/3                                                                          
                        time:   [105.37 ms 106.75 ms 107.54 ms]
                        thrpt:  [27.897  elem/s 28.103  elem/s 28.470  elem/s]
                 change:
                        time:   [-7.3565% -6.0757% -4.8199%] (p = 0.00 < 0.05)
                        thrpt:  [+5.0640% +6.4687% +7.9406%]
                        Performance has improved.
```

I wonder if one of our baselines is off?