@mpapierski 2020-01-07
> 

Maybe its my machine, although I confirmed it again with new dev (328b1d7f5cb8bd9405ceca8ad67d484f16522a51) and I'm seeing similar gains