@mpapierski 2020-01-10
Tried this again recently so it might be my machines problem, but I consistently see performance improvements compared to dev baseline similar to mentioned in the first comment (even with newer dev than this branch root). Anyway glad others also experience performance improvements!